net/tap: do not duplicate file descriptors
The TAP device can use same file descriptor for both Rx and Tx queues
which reduces the number of FD's required.

MP process support passes file descriptors from primary
to secondary process; but because of the restriction on
max FD's passed RTE_MP_MAX_FD_NUM (8) the TAP device was restricted
to only 4 queues if using secondary.
This allows up to 8 queues (versus 4).

The restriction on max FD's should be changed in EAL in
future, but it will break ABI compatibility.
The max Linux supports which is SCM_MAX_FD (253).

Signed-off-by: Stephen Hemminger <stephen@networkplumber.org>
Acked-by: Ferruh Yigit <ferruh.yigit@amd.com>