net/tap: do not build flow support if header is out of date
The proper place for finding bpf structures and functions is
in linux/bpf.h. The original version was trying to workaround the
case where the build environment was running on old pre BPF
version of Glibc, but the target environment had BPF.

Having own private (and divergent) version headers leads to future
problems when BPF definitions evolve.

If the build infrastructure is so old that TC flower is not
supported, then the TAP device will build without any flow support.

Signed-off-by: Stephen Hemminger <stephen@networkplumber.org>
Acked-by: Ferruh Yigit <ferruh.yigit@amd.com>